jakaa sovelluksen moduuleihin
