+++ index.js
@@ -1,47 +1,28 @@
-const http = require('http')
 const express = require('express')
 const app = express()
 const bodyParser = require('body-parser')
 const cors = require('cors')
 const mongoose = require('mongoose')
 const morgan = require('morgan')
+const blogsRouter = require('./controllers/blogs')
 
 if ( process.env.NODE_ENV !== 'production' ) {
   require('dotenv').config()
 }
 
-const Blog = mongoose.model('Blog', {
-  title: String,
-  author: String,
-  url: String,
-  likes: Number
-})
-
-module.exports = Blog
-
 app.use(cors())
 app.use(bodyParser.json())
 app.use(morgan('tiny'))
 
+app.use('/api/blogs', blogsRouter)
+
 const mongoUrl = process.env.MONGODB_URI
 mongoose.connect(mongoUrl, { useNewUrlParser: true })
-
-app.get('/api/blogs', (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+.then( () => {
+  console.log('Connected to database', mongoUrl)
 })
-
-app.post('/api/blogs', (request, response) => {
-  const blog = new Blog(request.body)
-
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
+.catch( error => {
+  console.log(error)
 })
 
 const PORT = 3003
